requirements.txt
livekit-agents[openai]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
pandas
openpyxl 